stream_socket_accept(), timeout support
- fixes https://github.com/peachpiecompiler/peachpie/issues/916
- timeout is used and function returns FALSE with a warning in case of timeout
- timeout is float

+++ src/Peachpie.Runtime/Context.Configuration.cs
@@ -159,7 +159,7 @@ namespace Pchp.Core
         /// <summary>
         /// Default timeout for socket based streams.
         /// </summary>
-        public int DefaultSocketTimeout = 60;
+        public double DefaultSocketTimeout = 60.0;
 
         /// <summary>
         /// A default file open mode used when it is not specified in <c>fopen</c> function explicitly.
